[refactor] 파일 디렉터리 구조 통일
- 각 페이지의 진입점이 되는 컴포넌트를 pages 디렉터리에 두고 그 안에서 필요한 컴포넌트들을 components 디렉터리에 두기로 결정했습니다.
