+++ FE/src/pages/Login/GithubLoginProcess/GithubLoginProcess.scss
